Refatoração de códigos e estruturas.
