+++ data/sellers.txt
@@ -4,3 +4,4 @@ boneca;23-232323222;boneca@olist.com
 boneca;23-232323221;boneca@olist.com
 eu;11-999999999;contato@contato.com
 Ralph;35-123456789;gustavo.sobreiro@olist.com
+Gustavo;35-999999999;contato@contato.com
